+++ src/main/java/deokhugam/deokhugam/user/service/BasicUserService.java
@@ -7,6 +7,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import deokhugam.deokhugam.user.entity.User;
 import deokhugam.deokhugam.user.exception.UserNotFoundException;
+import deokhugam.deokhugam.user.exception.UserPermissionException;
 import deokhugam.deokhugam.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 
@@ -33,26 +34,35 @@ public class BasicUserService implements UserService {
 
 	@Override
 	@Transactional
-	public User update(UUID userId, String nickname) {
-		User user = userRepository.findById(userId)
-			.orElseThrow(() -> UserNotFoundException.withId(userId));
+	public User update(UUID pathUserId, UUID loginUserId, String nickname) {
+		if (!pathUserId.equals(loginUserId)) {
+			throw UserPermissionException.withPathUserIdAndLoginUserId(pathUserId, loginUserId);
+		}
+		User user = userRepository.findById(pathUserId)
+			.orElseThrow(() -> UserNotFoundException.withId(pathUserId));
 		user.updateProfile(nickname);
 		return user;
 	}
 
 	@Override
 	@Transactional
-	public void softDelete(UUID userId) {
-		User user = userRepository.findById(userId)
-			.orElseThrow(() -> UserNotFoundException.withId(userId));
+	public void softDelete(UUID pathUserId, UUID loginUserId) {
+		if (!pathUserId.equals(loginUserId)) {
+			throw UserPermissionException.withPathUserIdAndLoginUserId(pathUserId, loginUserId);
+		}
+		User user = userRepository.findById(pathUserId)
+			.orElseThrow(() -> UserNotFoundException.withId(pathUserId));
 		user.softDelete();
 	}
 
 	@Override
 	@Transactional
-	public void hardDelete(UUID userId) {
-		User user = userRepository.findById(userId)
-			.orElseThrow(() -> UserNotFoundException.withId(userId));
+	public void hardDelete(UUID pathUserId, UUID loginUserId) {
+		if (!pathUserId.equals(loginUserId)) {
+			throw UserPermissionException.withPathUserIdAndLoginUserId(pathUserId, loginUserId);
+		}
+		User user = userRepository.findById(pathUserId)
+			.orElseThrow(() -> UserNotFoundException.withId(pathUserId));
 		userRepository.delete(user);
 	}
 }
